LayoutTests/fast/dom/DOMURL/url-username-expected.txt

Test getting the username attribute of a URL object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Default username is empty
PASS a.username is ''
Unspecified username should return empty string
PASS a.username is ''
Empty username should be empty
PASS a.username is ''
Regular username retrieval
PASS a.username is 'joebob1'
Username with empty password
PASS a.username is 'joebob2'
Username with non-empty password
PASS a.username is 'joebob3'
Setting username
PASS a.href is 'http://joebob4@example.com/'
PASS successfullyParsed is true

TEST COMPLETE
